+++ libs/search-widget/src/components/label/Label.svelte
